Implement subgraph solvin
Description:
Once the constraint graph is decomposed into smaller triangles it can be
easily solved by quadratic equations.  The solver has to solve each
triangle and reconstruct the gcs

Solution:
Implemented a defaultSolver that uses that strategy of starting from a
zerod triangle, then  building up from there every part that has at
least two solved parts. The algorithm is very primitive and needs
refactoring, however it serves as a working example

+++ src/constraint_solver/constraints.hpp
@@ -13,8 +13,8 @@ concept ConstraintType = requires(T constraint) {
 };
 
 struct DistanceConstraint {
-    float distance;
-    explicit DistanceConstraint(float d)
+    double distance;
+    explicit DistanceConstraint(double d)
         : distance { d }
     {
     }
@@ -22,8 +22,8 @@ struct DistanceConstraint {
     double getConstraintValue() const { return distance; }
 };
 struct TangencyConstraint {
-    float angle;
-    explicit TangencyConstraint(float d)
+    double angle;
+    explicit TangencyConstraint(double d)
         : angle { d }
     {
     }
